Compiled code & ran a SupportSystem object
Adding a print statement with nothing in the parenthesis will cause an error to be present, since the Responder class only takes statements under put instead of print.  Because of that, placing \n AFTER something like crash, as seen in this commit, the terminal window got the result I was looking for.